% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmetric.R
\name{compute_union}
\alias{compute_union}
\title{Helper function. Compute the area of the union of the identified geometries.}
\usage{
compute_union(ref_sf, seg_sf, ref_id, seg_id, id_ref, id_seg)
}
\arguments{
\item{ref_sf}{An sf object. The reference.}

\item{seg_sf}{An sf object. The segmentation.}

\item{ref_id}{An integer. Name of the Id field in the reference object.}

\item{seg_id}{An integer. Name of the Id field in the segmentation object.}

\item{id_ref}{An integer. Id of a reference geometry.}

\item{id_seg}{An integer. Id of a segmentation geometry.}
}
\value{
\preformatted{  A numeric.
}
}
\description{
Helper function. Compute the area of the union of the identified geometries.
}
