% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmetric.R
\name{get_metrics}
\alias{get_metrics}
\title{Gather the required information for computing the metrics.}
\usage{
get_metrics(ref_sf, seg_sf, ref_id, seg_id, alpha = 0.5)
}
\arguments{
\item{ref_sf}{An sf object. The reference segmentation.}

\item{seg_sf}{An sf object. The segmentation under test.}

\item{ref_id}{A character. Name of the field containing unique Ids.}

\item{seg_id}{A character. Name of the field containing unique Ids.}

\item{alpha}{A length-one numeric. Weight used to adjust the sensitivity of
the F_measure to Over and Under Segmentation.}
}
\value{
A data.frame
\itemize{
\item ref_id - ID of the reference polygons.
\item seg_id - ID of the segmentation polygons.
\item intersect_area - Area of the intersection between the reference and segmentation polygons.
\item union_area - Area of the union between the reference and the segmentation polygons.
\item ref_area - Area of the reference polygons.
\item seg_area - Area of the segmentation polygons.
\item centroid_in_ref - Are the centroids of the segmentation polygons inside the reference?
\item rate_int_ref_area - Proportion of the intersection to the reference polygon area.
\item rate_int_seg_area - Proportion of the intersection to the segmentation polygons area.
\item jaccard - Jaccard index.
\item oseg - Oversegmentation index.
\item useg - Undersegmentation index.
\item afi - Area Fit Index (AFI) index.
}
}
\description{
Gather the required information for computing the metrics.
}
