% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs}{left hand side parameter.}

\item{rhs}{right hand side parameter, usually a function.}
}
\value{
Return any object evaluated by function call \code{rhs(lhs, ...)}.
}
\description{
Imported function from \code{magrittr}.
}
