% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_sf}
\alias{ref_sf}
\alias{sample_ref_sf}
\title{LEM+ dataset}
\format{
These datasets are objects of class \code{sf} (inherited from \code{tbl_df},
\code{tbl}, \code{data.frame}) with 2 variables:
\itemize{
\item{\code{id}: identification}
\item{\code{geometry}: polygons}
}

\code{ref_sf}: a dataset with 195 features.

\code{sample_ref_sf}: a dataset with 5 features.
}
\source{
Oldoni et al. (2020) \doi{10.1016/j.dib.2020.106553}.
}
\usage{
ref_sf

sample_ref_sf
}
\description{
\code{ref_sf}: a dataset containing field boundaries from Luiz Eduardo Magalhaes
municipality, Brazil.

The data covers the following extent:
xmin: -46.37683 ymin: -12.34579 xmax: -46.15776 ymax: -12.13663
CRS: EPSG:4326

\code{sample_ref_sf}: a subset of \code{ref_sf} dataset.
}
\examples{
data("ref_sf", package = "segmetric")
data("sample_ref_sf", package = "segmetric")
}
\references{
\itemize{
\item Oldoni, L.V., Sanches, I.D.A., Picoli, M.C.A.,
Covre, R.M. and Fronza, J.G., 2020. LEM+ dataset: For
agricultural remote sensing applications.
Data in Brief, 33, p.106553.
}
}
\keyword{datasets}
