% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_functions}
\alias{db_functions}
\alias{sm_list_metrics}
\alias{sm_new_metric}
\alias{sm_reg_metric}
\alias{sm_unreg_metric}
\alias{sm_desc_metric}
\title{Database of metrics functions}
\usage{
sm_list_metrics()

sm_new_metric(fn, name = "", description = "", reference = "")

sm_reg_metric(metric_id, entry)

sm_unreg_metric(metric_id)

sm_desc_metric(metric_id)
}
\arguments{
\item{fn}{A \code{function} that receives a \code{segmetric} object and
returns the metric values.}

\item{name}{A \code{character} containing the metric name
(Defaults to \code{""}).}

\item{description}{A \code{character} containing a description of the metric
(Defaults to \code{""}).}

\item{reference}{A \code{character} with the reference to a scientific
literature describing the metric.}

\item{metric_id}{A \code{character} value containing a unique metric
identification.}

\item{entry}{A \code{metric_entry} object returned by \code{new_metric()}
function.}
}
\description{
These functions are used to register new metrics in the \code{segmetric}
database.
\itemize{
\item \code{sm_list_metrics()} list all registered metrics.
\item \code{sm_new_metric()} create a new metric entry to be registered in
the database.
\item \code{sm_reg_metric()} register a new metric in the database.
\item \code{sm_unreg_metric()} remove a metric entry from the database.
\item \code{sm_desc_metric()} describe a metric registered from the database.
}
}
\examples{
sm_reg_metric(
    metric_id = "Example",
    entry = sm_new_metric(
        fn = function(m) {
            sm_area(sm_ytilde(m)) / 
                sm_area(sm_ref(m), order = sm_ytilde(m))
        },
        name = "Metric name example",
        description = paste("Values range from A to B.",
                            "Optimal value is C"),
        reference = "Author (Year)"
    ))

sm_desc_metric("Example")
sm_list_metrics()

}
