% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{general_functions}
\alias{general_functions}
\alias{sm_area}
\alias{sm_centroid}
\alias{sm_intersection}
\alias{sm_subset_union}
\alias{sm_rbind}
\alias{sm_norm_left}
\alias{sm_norm_frac}
\title{General functions}
\usage{
sm_area(s, order = NULL)

sm_centroid(s)

sm_intersection(s1, s2, touches = TRUE)

sm_subset_union(s)

sm_rbind(...)

sm_norm_left(x, y)

sm_norm_frac(x, y)
}
\arguments{
\item{s}{A \code{sf} object. Either a reference, a segmentation, or a subset.}

\item{order}{A \code{sf} object of type subset\code{sf}.}

\item{s1}{A \code{sf} object. Either a reference, a segmentation, or a subset.}

\item{s2}{A \code{sf} object. Either a reference, a segmentation, or a subset.}

\item{touches}{A logical. Is the border part of the intersection?}

\item{...}{Set of \code{sf} objects of type subset\code{sf}.}

\item{x}{A numerator of a fraction.}

\item{y}{A denominator of a fraction.}
}
\description{
These functions manipulate segmetric objects.
\itemize{
\item \code{sm_area()} ...
\item \code{sm_centroid()} ...
\item \code{sm_intersection()} ...
\item \code{sm_subset_union()} ...
\item \code{sm_rbind()} ...
\item \code{norm_left()} ...
\item \code{norm_frac()} ...
\item \code{norm_diff()} ...
}
}
