% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{general_functions}
\alias{general_functions}
\alias{sm_area}
\alias{sm_centroid}
\alias{sm_intersections}
\alias{sm_union}
\alias{rbind_distinct}
\alias{sm_norm_left}
\alias{sm_norm_frac}
\alias{norm_diff}
\title{General functions}
\usage{
sm_area(s, order = NULL)

sm_centroid(s)

sm_intersections(s1, s2, touches = TRUE)

sm_union(s1, s2, order)

rbind_distinct(...)

sm_norm_left(x, y)

sm_norm_frac(x, y)

norm_diff(x, y)
}
\arguments{
\item{s}{A \code{sf} object. Either a reference, a segmentation, or a subset.}

\item{order}{A \code{sf} object of type subset\code{sf}. This argument is equivalent
to left join. The return of the function are ordered according to the object
passed to this parameter.}

\item{s1}{A \code{sf} object. Either a reference, a segmentation, or a subset.}

\item{s2}{A \code{sf} object. Either a reference, a segmentation, or a subset.}

\item{touches}{A logical. Is the border part of the intersection?}

\item{...}{Set of \code{sf} objects of type subset\code{sf}.}

\item{x}{A numerator of a fraction.}

\item{y}{A denominator of a fraction.}
}
\description{
These functions manipulate segmetric objects.
\itemize{
\item \code{sm_area()} Return a vector of areas, one for each polygon.
\item \code{sm_centroid()} Return the centroids of the given polygons.
\item \code{sm_intersections()} Return the intersection of the given simple features.
\item \code{sm_union()} Return the unio of the given simple features.
\item \code{rbind_distinct()} Return the merge of unique simple features.
\item \code{sm_norm_left()} Return the normalized value by x.
\item \code{sm_norm_frac()} Return the normalized value by y.
}
}
