% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{general_functions}
\alias{general_functions}
\alias{sm_area}
\alias{sm_centroid}
\alias{sm_intersection}
\alias{sm_subset_union}
\alias{sm_rbind}
\alias{sm_group_by}
\title{General functions}
\usage{
sm_area(s, order = NULL)

sm_centroid(s)

sm_intersection(s1, s2, touches = TRUE)

sm_subset_union(s)

sm_rbind(...)

sm_group_by(s, by, fn, ...)
}
\arguments{
\item{s, s1, s2}{Either a \code{ref_sf}, a \code{seg_sf}, or a \code{subset_sf} object
(inherits from \code{sf}).}

\item{order}{A \code{subset_sf}. This argument arranges the returned values
according to the object passed here.}

\item{touches}{A \code{logical}. Is the border part of the intersection?}

\item{...}{For \code{sm_rbind()}, a set of \code{sf} objects of type subset \code{sf}.
For \code{sm_group_by()}, extra parameter to pass to \code{fn} function.}

\item{by}{A \code{character} value with the column to group.}

\item{fn}{A \code{function} to apply on each group.}
}
\description{
These functions manipulate segmetric objects.
\itemize{
\item \code{sm_area()} returns a vector of areas, one for each polygon.
\item \code{sm_centroid()} returns the centroids of the given polygons.
\item \code{sm_intersection()} returns the intersection of the given simple features.
\item \code{sm_subset_union()} returns the union of the given simple features.
\item \code{sm_rbind()} returns the merge of unique simple features.
\item \code{sm_group_by()} apply a function to groups of \code{subset_sf}.
}
}
