% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-funs.R
\name{metric_functions}
\alias{metric_functions}
\alias{sm_compute}
\alias{OS1}
\alias{US1}
\alias{OS2}
\alias{US2}
\alias{OS3}
\alias{US3}
\alias{AFI}
\alias{QR}
\alias{D_index}
\alias{precision}
\alias{recall}
\alias{UMerging}
\alias{OMerging}
\alias{M}
\alias{E}
\alias{RAsub}
\alias{RAsuper}
\alias{PI}
\alias{Fitness}
\alias{ED3}
\alias{F_measure}
\title{Metric functions}
\usage{
sm_compute(m, metric_id, ...)

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "US1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "OS1")

sm_compute(m, metric_id = "F_measure", alpha = 0.5)
}
\arguments{
\item{m}{A \code{segmetric} object.}

\item{metric_id}{A character value with the name of a metric to be computed.}

\item{...}{Any additional argument to compute
a metric (see details).}
}
\description{
\code{sm_compute()} is used to compute a given metric (\code{metric_id} parameter)
from \code{segmetric} objects. It compares the reference to the segmentation
polygons using a metric.

A list with all supported metrics can be obtained
by \code{sm_list_metrics()} (see details for more information).
}
\details{
\code{OS1}: refers to Over-segmentation. Values range .....

\code{"US1"} refers to Under-segmentation. Values range .....
(Costa et al., 2010)

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{OS1}: refers to Over-segmentation. Values range .....

\code{"F_measure"}: refers to F-measure metric (Van Rijsbergen, 1979;
Zhang et al., 2015). Values range from 0 to 1 (optimal). It takes the
optional weight argument \code{alpha}, which ranges from \code{0.0} to \code{1.0}
(default is \code{0.5}).
}
\examples{
# load sample datasets
data("sample_ref_sf", package = "segmetric")
data("sample_seg_sf", package = "segmetric")

# create segmetric object
m <- sm_read(ref_sf = sample_ref_sf, seg_sf = sample_seg_sf)

# compute AFI metric and summarize it
sm_compute(m, "AFI") \%>\% summary()

# compute OS1, F_measure, and US2 metrics
m <- sm_compute(m, "OS1") \%>\%
  sm_compute("F_measure") \%>\%
  sm_compute("US2")

# summarize them
summary(m)

}
\references{
A complete list of cited references is available at \code{?segmetric}.
}
\seealso{
\code{sm_list_metrics()}
}
