% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-funs.R
\name{metric_functions}
\alias{metric_functions}
\alias{sm_compute}
\alias{OS1}
\alias{US1}
\alias{OS2}
\alias{US2}
\alias{OS3}
\alias{US3}
\alias{AFI}
\alias{QR}
\alias{D_index}
\alias{precision}
\alias{recall}
\alias{UMerging}
\alias{OMerging}
\alias{M}
\alias{E}
\alias{RAsub}
\alias{RAsuper}
\alias{PI}
\alias{Fitness}
\alias{ED3}
\alias{F_measure}
\title{Metric functions}
\usage{
sm_compute(m, metric_id, ...)

OS1(m)

US1(m)

OS2(m)

US2(m)

OS3(m)

US3(m)

AFI(m)

QR(m)

D_index(m)

precision(m)

recall(m)

UMerging(m)

OMerging(m)

M(m)

E(m)

RAsub(m)

RAsuper(m)

PI(m)

Fitness(m)

ED3(m)

F_measure(m, alpha = 0.5)
}
\arguments{
\item{m}{A \code{segmetric}object.}

\item{metric_id}{A character value with the name of a metric to be computed.}

\item{...}{Any additional argument to compute
a metric (see details).}

\item{alpha}{Numeric value ranging from 0 to 1. See Details.}
}
\description{
\code{sm_compute()} is used to compute a given metric (\code{metric_id} parameter)
from \code{segmetric} objects. It compares the reference to the segmentation
polygons using a metric.

A list with all supported metrics can be obtained
by \code{sm_list_metrics()} (see details for more information).
}
\details{
\code{"OS1"} refers to Oversegmentation. Its values range from 0 (optimal) to 1
(Clinton et al., 2010).

\code{"US1"} refers to Undersegmentation. Its values range from 0 (optimal) to 1
(Clinton et al., 2010).

\code{"OS2"} refers to Oversegmentation. Its values range from 0 (optimal) to 1
(Persello and Bruzzone, 2010).

\code{"US2"} refers to Undersegmentation. Its values range 0 (optimal) to 1
(Persello and Bruzzone, 2010).

\code{"OS3"} refers to Oversegmentation. Its values range from 0 (optimal) to 1
(Yang et al., 2014).

\code{"US3"} refers to Undersegmentation. Its values range from 0 (optimal) to 1
(Yang et al., 2014).

\code{"AFI"} refers to Area Fit Index. Its optimal value is 0 (Lucieer and Stein,
2002; Clinton et al., 2010).

\code{"QR"} refers to Quality rate. Its values range from 0 (optimal) to 1
(Weidner, 2008; Clinton et al., 2010).

\code{"D_index"} refers to Index D. Its values range from 0 (optimal) to 1
(Levine and Nazif, 1982; Clinton et al., 2010).

\code{"precision"} refers to Precision. Its values range from 0 to 1 (optimal)
(Van Rijsbergen, 1979; Zhang et al., 2015).

\code{"recall"} refers to Recall. Its values range from 0 to 1 (optimal) (Van
Rijsbergen, 1979; Zhang et al., 2015).

\code{"UMerging"} refers to Undermerging. Its values range from 0 (optimal) to 0.5
(Levine and Nazif, 1982; Clinton et al., 2010).

\code{"OMerging"} refers to Overmerging. Its values range from 0 (optimal) to 0.5
(Levine and Nazif, 1982; Clinton et al., 2010).

\code{"M"} refers to Match. Its values range from 0 to 1 (optimal) (Janssen and
Molenaar, 1995; Feitosa et al., 2010).

\code{"E"} refers to Evaluation measure. Its values range from 0 (optimal) to 50
(Carleer et al., 2005).

\code{"RAsub"} refers to Relative area. Its values range from 0 to 1 (optimal)
(Müller et al., 2007; Clinton et al., 2010).

\code{"RAsuper"} refers to Relative area. Its values range from 0 to 1 (optimal)
(Müller et al., 2007; Clinton et al., 2010).

\code{"PI"} refers to Purity Index. Its values range from 0 to 1 (optimal) (van
Coillie et al., 2008).

\code{"Fitness"} refers to Fitness function. Its optimal value is 0 (Costa et al.,
2008).

\code{"ED3"} refers to Euclidean Distance. Its values range from 0 (optimal) to 1
(Yang et al., 2014).

\code{"F_measure"} refers to F-measure metric. Its values range from 0 to 1
(optimal) (Van Rijsbergen, 1979; Zhang et al., 2015). It takes the
optional weight argument \code{alpha}, which ranges from \code{0.0} to \code{1.0}
(the default is \code{0.5}).
}
\examples{
# load sample datasets
data("sample_ref_sf", package = "segmetric")
data("sample_seg_sf", package = "segmetric")

# create segmetric object
m <- sm_read(ref_sf = sample_ref_sf, seg_sf = sample_seg_sf)

# compute AFI metric and summarize it
sm_compute(m, "AFI") \%>\% summary()

# compute OS1, F_measure, and US2 metrics
m <- sm_compute(m, "OS1") \%>\%
  sm_compute("F_measure") \%>\%
  sm_compute("US2")

# summarize them
summary(m)

}
\references{
A complete list of cited references is available at \code{?segmetric}.
}
\seealso{
\code{sm_list_metrics()}
}
