% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{segmetric_functions}
\alias{segmetric_functions}
\alias{sm_read}
\alias{sm_clear}
\alias{sm_compute}
\alias{summary.segmetric}
\title{General functions}
\usage{
sm_read(ref_sf, seg_sf)

sm_clear(m)

sm_compute(m, metric_id, ...)

\method{summary}{segmetric}(m, weight = NULL, ...)
}
\arguments{
\item{ref_sf}{A \code{sf} object...}

\item{seg_sf}{A \code{sf} object...}

\item{m}{A \code{segmetric} object ...}

\item{...}{Additional parameters. See details.}

\item{weight}{A \code{numeric} vector ...}

\item{metric}{A \code{character} value ...}
}
\description{
These functions manipulate segmetric objects.
\itemize{
\item \code{sm_read()} ...
\item \code{sm_clear()} ...
\item \code{sm_compute()} ...
\item \code{summary.segmetric()} ...
\item \code{sm_ref_area()} ...
\item \code{sm_seg_area()} ...
\item \code{sm_inter_area()} ...
}
}
\examples{

}
