% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-funs.R
\name{set_functions}
\alias{set_functions}
\alias{sm_ytilde}
\alias{sm_xtilde}
\alias{sm_yprime}
\alias{sm_xprime}
\alias{sm_ya}
\alias{sm_yb}
\alias{sm_yc}
\alias{sm_yd}
\alias{sm_ystar}
\alias{sm_ycd}
\alias{sm_ye}
\alias{sm_yf}
\alias{sm_yg}
\title{Set functions}
\usage{
sm_ytilde(m)

sm_xtilde(m)

sm_yprime(m)

sm_xprime(m)

sm_ya(m)

sm_yb(m)

sm_yc(m)

sm_yd(m)

sm_ystar(m)

sm_ycd(m)

sm_ye(m)

sm_yf(m)

sm_yg(m)
}
\arguments{
\item{m}{A \code{segmetric} object.}
}
\description{
These functions manipulate segmetric objects.
\itemize{
\item \code{sm_ref()} Get the set of n polygons of reference, represented by \eqn{X = \{x_{i}: i = 1, ....., n\}}
\item \code{sm_seg()} Get the set of m segmentation polygons, represented by \eqn{Y = \{y_{j}: j = 1, ....., m\}}
\item \code{sm_ytilde()} Get the \eqn{Y} subset \eqn{\tile{Y}_{i} = \{y_{j}: area(x_{i} \cap y_{j}) \neq 0\}}
\item \code{sm_xtilde()} Get the \eqn{X} subset \eqn{\tile{X}_{j} = \{x_{i}: area(y_{j} \cap x_{i}) \neq 0\}}
\item \code{sm_yprime()} Get the \eqn{Y} subset \eqn{Y'_{i} = \{y_{j}: max(area(x_{i} \cap y_{j}))\}}
\item \code{sm_xprime()} Get the \eqn{X} subset \eqn{X'_{j} = \{x_{i}: max(area(y_{j} \cap x_{i}))\}}
\item \code{sm_ya()} Get the \eqn{\tilde{Y}_{i}$ subset $Ya_{i} = \{y_{j}: the centroid of x_{i} is in y_{j}\}}
\item \code{sm_yb()} Get the \eqn{\tilde{Y}_{i}$ subset $Yb_{i} = \{y_{j}: the centroid of y_{j} is in x_{i}\}}
\item \code{sm_yc()} Get the \eqn{\tilde{Y}_{i}$ subset $Yc_{i} = \{y_{j}: area(x_{i} \cap y_{j}) \div area(y_{j}) > 0.5\}}
\item \code{sm_yd()} Get the \eqn{\tilde{Y}_{i}$ subset $Yd_{i} = \{y_{j}: area(x_{i} \cap y_{j}) \div area(x_{i}) > 0.5\}}
\item \code{sm_ystar()} Get the subset \eqn{\star{Y}_{i} = Ya_{i} \cup Yb_{i} \cup Yc_{i} \cup Yc_{i}}
\item \code{sm_ycd()} Get the subset \eqn{Ycd_{i} = Yc_{i} \cup Yd_{i}}
\item \code{sm_ye()} Get the \eqn{\tilde{Y}_{i}$ subset $Ye_{i} = \{y_{j}: area(x_{i} \cap y_{j}) \div area(y_{j}) = 1\}}
\item \code{sm_yf()} Get the \eqn{\tilde{Y}_{i}$ subset $Yf_{ì} = \{y_{j}: area(x_{i} \cap y_{j}) \div area(y_{j}) > 0.55\}} 
\item \code{sm_yg()} Get the \eqn{\tilde{Y}_{i}$ subset $Yg_{ì} = \{y_{j}: area(x_{i} \cap y_{j}) \div area(y_{j}) > 0.75\}} 
}
}
\references{
Clinton, N., Holt, A., Scarborough, J., Yan, L., & Gong, P. (2010). Accuracy Assessment Measures for Object-based Image Segmentation Goodness. Photogrammetric Engineering & Remote Sensing, 76(3), 289–299. https://doi.org/10.14358/PERS.76.3.289
Costa, H., Foody, G. M., & Boyd, D. S. (2018). Supervised methods of image segmentation accuracy assessment in land cover mapping. Remote Sensing of Environment, 205(December 2017), 338–351. https://doi.org/10.1016/j.rse.2017.11.024
}
