% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_handling_functions}
\alias{subset_handling_functions}
\title{Subset handling functions}
\arguments{
\item{s}{A \code{sf} object...}

\item{allowed_types}{...}

\item{...}{...}

\item{m}{A \code{segmetric} object ...}

\item{subset_id}{A \code{character} value ...}

\item{expr}{A \code{numeric} vector ...}

\item{inset}{...}
}
\description{
These functions manipulate segmetric objects.
\itemize{
\item \code{sm_list_subsets()} ...
\item \code{sm_eval_subset()} ...
\item \code{sm_get_subset()} ...
}
}
\examples{

}
